requirements.txt
alembic==1.7.5
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
gunicorn==20.1.0
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
requests==2.27.1
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
